requirements.txt
chromadb
ollama
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
